+++ src/modules/WindowLogin/index.ts
@@ -1,0 +1,3 @@
+import WindowLogin from "./WindowLogin";
+
+export default WindowLogin;
