+++ src/modules/WindowLogin/logic/closeWindow.ts
@@ -1,0 +1,4 @@
+export default function closeWindow() {
+  const blur = document.getElementById("blur") as Element;
+  blur.innerHTML = "";
+}
