+++ src/modules/WindowLogin/logic/openWindow.ts
@@ -1,0 +1,7 @@
+import { createApp } from "@/ourReact/jsx-runtime";
+import WindowLogin from "../WindowLogin";
+
+export default function openWindow() {
+  const blur = document.getElementById("blur") as Element;
+  createApp(blur, WindowLogin);
+}
